Moved utility funtions to utils.py
